+++ src/Routemeister.UnitTests/Routemeister.UnitTests.csproj
@@ -59,9 +59,9 @@
     <Compile Include="MessageEnvelopeTests.cs" />
     <Compile Include="MessageRouteFactoryTests.cs" />
     <Compile Include="MessageRoutesTests.cs" />
-    <Compile Include="MiddlewareEnabledSequentialAsyncRouterTests.cs" />
+    <Compile Include="Routers\MiddlewareEnabledAsyncRouterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SequentialAsyncRouterTests.cs" />
+    <Compile Include="Routers\SequentialAsyncRouterTests.cs" />
     <Compile Include="UnitTestsOf.cs" />
   </ItemGroup>
   <ItemGroup>
